function to replace ezSlider added
problem with EZSlider is that its values and actionValues are all post
control spec. What I really want is that everything is 0 to 1 values
for as long as possible. Anyway, here is the function and the next
thing is to replace the ezSliders with this function and get that to
work.